% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mag_for_pow.R
\name{abs_max}
\alias{abs_max}
\title{Helper function for est_local_pow.  This function
is used to estimate the magnitude needed for a
certain direction to achive a given power.}
\usage{
abs_max(x)
}
\arguments{
\item{boot_data}{bootstrap estimates of the (centered) limiting distribution.
Data where columns correspond to different covariates, and rows are independent observations.}

\item{lp}{the lp norm to be used (ideally it is an integer)}

\item{dir}{the direction for which we wish to find the magnitude needed to achieve
the speceified power.}

\item{power}{power which the local alternative is to achieve
for a magnitude that will be solved.}
}
\value{
Magnitude for the specified lp norm for a given local alternative.
}
\description{
Helper function for est_local_pow.  This function
is used to estimate the magnitude needed for a
certain direction to achive a given power.
}
