% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amar.R, R/vol.R
\name{predict.amar.fit}
\alias{amar}
\alias{amar.design.matrix}
\alias{amar.response}
\alias{amar.sim}
\alias{amar.train}
\alias{predict.amar.fit}
\alias{vol.exp.smoothing}
\title{Predict AMAR object}
\usage{
\method{predict}{amar.fit}(object, newx)

amar(x, max.order, method = c("bic", "threshold"), threshold,
  threshold.const = 0.5, max.scales = 10, verbose = TRUE,
  use.RcppEigen = TRUE, ...)

amar.sim(n, coefficients, scales, sigma = 1)

amar.train(x, max.order, split.pct = c(0.5, 0.25, 0.25),
  remove.volatility = TRUE, volatility.criterion = c("jarque.bera"),
  target = c("hit.rate", "r.squared"), max.scales = 10, verbose = TRUE,
  ...)

amar.design.matrix(x, scales)

amar.response(returns, scales)

vol.exp.smoothing(x, lambda)
}
\description{
Predict AMAR object

Amar fit

Simulate AMAR model

Train AMAR model

Create AMAR design matrix

Create AMAR response

Volatility Exponential Smoothing
}

