% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/answer2reviewers.R
\name{answer2reviewers}
\alias{answer2reviewers}
\title{answer2reviewers}
\usage{
answer2reviewers(reviewerID = 1)
}
\arguments{
\item{reviewerID}{an integer ranging from 1 to 3.}
}
\value{
an adequate answer
}
\description{
Generate an answer for your reviewers.
}
