% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amar.R, R/ar.R
\name{amar.fit}
\alias{amar.fit}
\alias{ar.design.matrix}
\alias{ar.fit}
\alias{ar.response}
\alias{ar.train}
\alias{predict.ar.fit}
\title{Fit AMAR.
Fit AMAR time series model.}
\usage{
amar.fit(x, scales, ...)

ar.fit(x, p, use.RcppEigen = TRUE, RcppEigen.method = 1, ...)

\method{predict}{ar.fit}(object, newx)

ar.train(x, max.order, split.pct = c(0.5, 0.25, 0.25),
  remove.volatility = TRUE, target = c("hit.rate", "r.squared"),
  verbose = TRUE, ...)

ar.design.matrix(returns, order)

ar.response(returns, order)
}
\arguments{
\item{x}{A numeric vector with data points.}

\item{...}{Currently not in use.}

\item{p}{Order of model to fit.}

\item{use.RcppEigen}{A logical variable. If \code{use.RcppEigen = TRUE}, the RcppEigen package is used to perform OLS estimation (recommended).}

\item{RcppEigen.method}{An integer specifining method used to solve OLS, see the RcppEigen package.}

\item{object}{An object returned by \link{\code{ar.fit}}.}

\item{newx}{A design matrix with p columns to be used for prediction.}

\item{x}{A numeric vector with data points.}

\item{...}{Other parameters that can be passed to \code{lm.fit} (If \code{use.RcppEigen = FALSE}).}
}
\value{
An object of class "amar.fit", which contains the following fields:
\item{response}{The response vector.}
\item{p}{Order of the fitted model.}
\item{coefficient}{OLS estimates of the AMAR coefficients.}
\item{fitted}{Fitted values.}
\item{residuals}{The difference between the response and the fitted value}

An object of class "ar.fit", which contains the following fields:
\item{response}{The response vector.}
\item{p}{Order of the fitted model.}
\item{coefficient}{OLS estimates of the autoregressive coefficients.}
\item{fitted}{Fitted values.}
\item{residuals}{The difference between the response and the fitted value}

Vector with the predicted values.
}
\description{
Fit AMAR.
Fit AMAR time series model.

Fit autoregressive with OLS
Fit an autoregressive time series model to the data using the Ordinary Least Squares method to estimate the autoregressive coefficients.

Predict \code{ar.fit} object
\code{predict} is a function to obtain predictions from  an \link{\code{ar.fit}} object.

Train AR model.

blablabl

blablabl
}

