% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laber_test.R
\name{bonf_test}
\alias{bonf_test}
\title{Run a bonferoni based test}
\usage{
bonf_test(obs_data, test_type = "pearson")
}
\arguments{
\item{obs_data}{The observed data}

\item{test_type}{Choose between using pearson correlation
or the mean as the parameter or intrest}
}
\description{
Run a bonferoni based test
}
