% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_test_stat.R
\name{calc_test_stat}
\alias{calc_test_stat}
\title{Helper function for cv_test.  This function estimates the cross validated
test statistic for a single fold of data.  Optionally this function can
also ouput the chosen norm index chosen using the training set.}
\usage{
calc_test_stat(
  lm_dst_est,
  obs_data,
  est_func,
  test_stat_func = l_p_norm,
  perf_meas,
  null_quants,
  norms_indx,
  norm_type,
  in_sq_mat = NULL
)
}
\arguments{
\item{lm_dst_est}{The limiting distribution of the vector of parameter estimates. Each row is an
observation, and each column corresponds to a parameter.}

\item{obs_data}{The observed data to be used for finding the optimal norm (training),
and finding the test statistic (testing).  Similar to above, each row is an observation and each
column corresponds to either the outcome (first column) or a covariate.}

\item{est_func}{A function that will provide the vector of parameter estimates for a subset of
obs_data.}

\item{test_stat_func}{A function that will provide the test statistic for the
given fold (using the testing data),
 and uses the best norm (decided on using the training data).}

\item{perf_meas}{The prefered measure to carry out evaluation of how far away the parameter
estimates are away from zero.}

\item{null_quants}{The 95 percent quantiles corresponding to the limiting distribution under the
various norms. The ordering of these quantiles is the same as that of the norm_indx.}

\item{norms_indx}{The index of the norms to be considered.  For example if we use the l_p norm,
norms_indx specifies the different p's to try.}

\item{norm_type}{The type of norm to be used for the test.  Generally the l_p norm}

\item{in_sq_mat}{Should either be NULL when normalization is not occuring, or should be the square root of the inverse of the estimated covariance matrix.}
}
\value{
learned test statistic for a single fold of data
}
\description{
Helper function for cv_test.  This function estimates the cross validated
test statistic for a single fold of data.  Optionally this function can
also ouput the chosen norm index chosen using the training set.
}
