% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cel_test.R
\name{cel_test}
\alias{cel_test}
\title{Runs a single test, with the posibility of using
either an apriori chosen l_p norm, or choosing
the norm adaptively}
\usage{
cel_test(obs_data, lp_norm = "ap", as_meas, num_boot = 1000,
  pos_lp = c(2, "max"), fixed = "power", cmp_meas = "mean")
}
\arguments{
\item{obs_data}{the observed data, with columns representing covariates and
rows representing observations.  The first column should be the outcome of interest}

\item{lp_norm}{the l_p norm to be used.  Select "ap" for adaptive selection of the
l_p norm.}

\item{as_meas}{the association measure to be used.}

\item{num_boot}{number of bootstrap samples used to approximate the limiting distribution}

\item{fixed}{the parameter (either power or magnitude) to be held constant in the
l_p norm selection proceedure.}

\item{cmp_meas}{the comparitave measure to use to select l_p norm (either mean or
worst direction).}
}
\value{
Either the average, or minimum power for the specified
lp norm for all local alternatives.
}
\description{
Runs a single test, with the posibility of using
either an apriori chosen l_p norm, or choosing
the norm adaptively
}
