% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IC-function.R, R/IC-functionbasis.R, R/cmpr.r
\name{cmdCVFolds}
\alias{cmdCVFolds}
\title{Create cross-validation folds}
\usage{
cmdCVFolds(n, cvControl)

cmdCVFolds(n, cvControl)

cmdCVFolds(n, cvControl)
}
\arguments{
\item{n}{Number of observations.}

\item{cvControl}{Named list generated by \link{cmdSuperLearner.CV.control}.}
}
\description{
This function generates cross-validation folds.

This function generates cross-validation folds.

This function generates cross-validation folds.
}
