% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IC-functionbasis.R, R/cmpr.r
\name{convert_to_basis}
\alias{convert_to_basis}
\title{Functions for converting from an IC and estimate of a function to the
corresponding basis function and IC.}
\usage{
convert_to_basis(est, ic, gridvals, num_basis = 10)

convert_to_basis(est, ic, gridvals, num_basis = 10)
}
\arguments{
\item{est}{estimate of function}

\item{ic}{IC estimate}
}
\description{
Functions for converting from an IC and estimate of a function to the
corresponding basis function and IC.

Functions for converting from an IC and estimate of a function to the
corresponding basis function and IC.
}
