% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_cv.R
\name{est_cv}
\alias{est_cv}
\title{This is an intermediate function inbetween the cv_test and one_fold_est functions which
allows for specification of the measure of performance.}
\usage{
est_cv(
  obs_data,
  param_est,
  control,
  est_lm_distr = NULL,
  return_lmd = FALSE,
  view_IC = FALSE
)
}
\arguments{
\item{obs_data}{The observed data to be used for finding the optimal norm (training),
and finding the test statistic (testing).  Similar to above, each row is an observation and each
column corresponds to either the outcome (first column) or a covariate.}

\item{param_est}{A function used to estimate both they parameter of interest
and the IC of the corresponding estimator.}

\item{control}{A list providing control arguments for the function.}

\item{est_lm_distr}{An estimate of the limiting distribution
if it is provided.}

\item{return_lmd}{Boolean for whether to return the estimated
limiting distribution.}

\item{view_IC}{Boolean for whether to see summaries of the IC}
}
\value{
learned test statistic for a single fold of data
}
\description{
This is an intermediate function inbetween the cv_test and one_fold_est functions which
allows for specification of the measure of performance.
}
