% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IC-calc.R
\name{est_influence_pearson}
\alias{est_influence_pearson}
\title{The following function will take a set of observations, and return the
variable of interest (the variable for which pearson correlation is
calculated.}
\usage{
est_influence_pearson(observ, trans = "none")
}
\arguments{
\item{observ}{the observed data}

\item{trans}{If there is a transformation of the parameter of interest}
}
\description{
The following function will take a set of observations, and return the
variable of interest (the variable for which pearson correlation is
calculated.
}
