% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_lmt_dstr.R
\name{est_limit_distr}
\alias{est_limit_distr}
\title{Generate an estimate of the limiting distribution of
the vector covariate specific parameter estimates}
\usage{
est_limit_distr(data, param, n = 1000)
}
\arguments{
\item{data}{observed data where columns correspond to different covariates, and rows are independent observations.}

\item{param}{The parameter of interest for example spearman correlation.}

\item{n}{the number of monte carlo draws. Default is 1000.}
}
\value{
An n by p matrix in which each column corresponds to the
bootstrap parameter estimate for the corresponding covariate under
the hypothesis that all parameter estimates are zero.
}
\description{
Generate an estimate of the limiting distribution of
the vector covariate specific parameter estimates
}
