% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_local_power.R
\name{est_loc_pow}
\alias{est_loc_pow}
\title{Generate an estimate of the limiting distribution of
the vector covariate specific parameter estimates}
\usage{
est_loc_pow(boot_data, lp = "max", method = "average", tries,
  magnitude = 3)
}
\arguments{
\item{boot_data}{bootstrap estimates of the (centered) limiting distribution. Data where
columns correspond to different covariates, and rows are independent observations.}

\item{lp}{the lp norm to be used}

\item{method}{either "average" for an average local power over a sphere of local
alternatives, or "min" to get the minimum power over a sphere of local alternatives}

\item{tries}{the number of places to check local alternatives either for averages or minima}

\item{magnitude}{geometric distance away from zero of the local alternative}
}
\value{
Either the average, or minimum power for the specified
lp norm for all local alternatives.
}
\description{
Generate an estimate of the limiting distribution of
the vector covariate specific parameter estimates
}
