% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laber_test.R
\name{est_pows}
\alias{est_pows}
\title{Estiamte the power using the generated test statistic, and
estimated distribution of the test statistic. Helper
function for ZL and ZL_use_infl}
\usage{
est_pows(tr_lm_dstr, ts_vec)
}
\arguments{
\item{tr_lm_dstr}{limiting distribution of the test statistic}

\item{ts_vec}{Vector containing the test statsitic}
}
\description{
Estiamte the power using the generated test statistic, and
estimated distribution of the test statistic. Helper
function for ZL and ZL_use_infl
}
