% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\name{gen_exp_cop}
\alias{gen_exp_cop}
\title{Functions for creating data}
\usage{
gen_exp_cop(n, d, cov)
}
\arguments{
\item{n}{number of observations}

\item{d}{number of dimensions}

\item{cov}{desired covariance}
}
\description{
Functions for creating data
}
