% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\name{gen_norm}
\alias{gen_norm}
\title{Generate observations from a normal distribution}
\usage{
gen_norm(n, d, cov)
}
\arguments{
\item{n}{number of observations}

\item{d}{number of covariates}

\item{cov}{variance-covariance of the multivariate normal distribution.}
}
\description{
Generate observations from a normal distribution
}
