% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\name{gen_one_samp_md}
\alias{gen_one_samp_md}
\title{Generate data from a Poission random variable while still obtaining the desired
xy correlation for each xy match and obtain the desired xx correlation among all
x variables. The y and x will always be marginally dependent due to the xx correlation.}
\usage{
gen_one_samp_md(base, xycor, xxcor, num_cor, num_uncor)
}
\arguments{
\item{base}{Base rate that is desired}

\item{xycor}{the correlation between each correlated x and y}

\item{xxcor}{the correlation between the x's}

\item{num_cor}{Number of correlated X's}

\item{num_uncor}{Number of uncorrelated X's}
}
\description{
Generate data from a Poission random variable while still obtaining the desired
xy correlation for each xy match and obtain the desired xx correlation among all
x variables. The y and x will always be marginally dependent due to the xx correlation.
}
