% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\name{gen_pois_samp}
\alias{gen_pois_samp}
\title{Generate data from a poission random variable while still obtaining the desired
xy correlation for each xy match and obtain the desired xx correlation among all
x variables}
\usage{
gen_pois_samp(ss, base, xycor, xxcor, num_cor, num_uncor, marg_indep = TRUE)
}
\arguments{
\item{ss}{sample size.}

\item{base}{Base rate that is desired}

\item{xycor}{the correlation between each correlated x and y}

\item{xxcor}{the correlation between the x's}

\item{num_cor}{Number of correlated X's}

\item{num_uncor}{Number of uncorrelated X's}

\item{marg_indep}{Are the X-values independent from one another}
}
\description{
Generate data from a poission random variable while still obtaining the desired
xy correlation for each xy match and obtain the desired xx correlation among all
x variables
}
