% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IC-calc.R
\name{get_infl}
\alias{get_infl}
\title{Wrapper function that provides all of the possible influence functions}
\usage{
get_infl(param, est_or_IC)
}
\arguments{
\item{param}{name of the parameter of interest}

\item{est_or_IC}{specify if an estimator of the parameter or IC is needed}
}
\description{
Wrapper function that provides all of the possible influence functions
}
