% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laber_test.R
\name{get_test_stat}
\alias{get_test_stat}
\title{Helper function for the Zhang and Laber tests (ZL and ZL_use_infl)}
\usage{
get_test_stat(obs_data)
}
\arguments{
\item{obs_data}{The observed data}
}
\description{
Helper function for the Zhang and Laber tests (ZL and ZL_use_infl)
}
