% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IC-function.R
\name{ic.func.est}
\alias{ic.func.est}
\title{Functional Null estimate based on the work of Ted Westling}
\usage{
ic.func.est(obs_dat, what = "both", control = NULL)
}
\arguments{
\item{obs_dat}{The observed data.}

\item{what}{should the IC, estimate, or both be returned.}

\item{control}{other control parameters.}
}
\description{
Functional Null estimate based on the work of Ted Westling
}
