% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IC-calc.R
\name{ic.pearson}
\alias{ic.pearson}
\title{The following function will take a set of observations, and return the
variable of interest (the variable for which pearson correlation is
calculated).}
\usage{
ic.pearson(observ, what = "both", control = NULL)
}
\arguments{
\item{observ}{the observed data.  The first column should be the outcome.}

\item{what}{the desired return value. Should be one of `"ic"`
(infludence curve), `"est"` (estimate), or `"both"`.}

\item{control}{any other control parameters to be passed to the estimator.}
}
\description{
The following function will take a set of observations, and return the
variable of interest (the variable for which pearson correlation is
calculated).
}
