% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IC-DE2.R
\name{ic.proj.rr}
\alias{ic.proj.rr}
\title{Estimate both the parameter, and the influence
curves used for estimating the projected risk ratio. The first column
of your data should correspond to the variable of interest.}
\usage{
ic.proj.rr(obs_data, what = "both", control = NULL)
}
\arguments{
\item{obs_data}{the observed data.  The first column should be the outcome.}

\item{what}{the desired return value. Should be one of `"ic"`
(infludence curve), `"est"` (estimate), or `"both"`.}

\item{control}{any other control parameters to be passed to the estimator.}
}
\description{
Estimate both the parameter, and the influence
curves used for estimating the projected risk ratio. The first column
of your data should correspond to the variable of interest.
}
