% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laber_test.R
\name{laber_test}
\alias{laber_test}
\title{Carry out the test described by Zhang and Laber}
\usage{
laber_test(obs_data, pos_lp_norms, n_mc_samples, nrm_type = "lp")
}
\arguments{
\item{obs_data}{The observed data}

\item{pos_lp_norms}{Potential norms to be considered}

\item{n_mc_samples}{the number of bootstrap samples to take when estimating the
limiting distribution}

\item{nrm_type}{the class of norms to be selected over.}
}
\description{
Carry out the test described by Zhang and Laber
}
