% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IC-DE2.R
\name{las_dw}
\alias{las_dw}
\title{Obtain an estimator of the probability delta = 1 given w}
\usage{
las_dw(delt_vec, obs_ws, simp = FALSE)
}
\arguments{
\item{delt_vec}{Vector of missingess indicators}

\item{obs_ws}{The observed w's}

\item{simp}{If the lasso function should simply return 0.5}
}
\description{
Obtain an estimator of the probability delta = 1 given w
}
