% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IC-DE2_nolas.R
\name{las_dw_nl}
\alias{las_dw_nl}
\title{Obtain an estimator of the probability delta = 1 given w}
\usage{
las_dw_nl(delt_vec, obs_ws)
}
\arguments{
\item{delt_vec}{Vector of missingess indicators}

\item{obs_ws}{The observed w's}
}
\description{
Obtain an estimator of the probability delta = 1 given w
}
