% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\name{make_data}
\alias{make_data}
\title{Generate data using one of the four specified models
from MCKEAGUE and QIAN paper}
\usage{
make_data(ss, dim, rho, model = 1, b = NULL)
}
\arguments{
\item{ss}{sample size (number of generated observations)}

\item{dim}{dimension of the generated data}

\item{rho}{between x correlation}

\item{model}{number indicating which model should be used}

\item{b}{local alternative to be used}
}
\description{
Generate data using one of the four specified models
from MCKEAGUE and QIAN paper
}
