% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no_fold_est.R
\name{no_fold_est}
\alias{no_fold_est}
\title{Helper function for cv_test.  This function estimates the cross validated
test statistic for a single fold of data.  Optionally this function can
also ouput the chosen norm index chosen using the training set.}
\usage{
no_fold_est(
  lm_dst_est,
  par_est,
  test_stat_func = l_p_norm,
  perf_meas,
  null_quants,
  norms_indx,
  norm_type
)
}
\arguments{
\item{lm_dst_est}{The limiting distribution of the vector of parameter estimates. Each row is an
observation, and each column corresponds to a parameter.}

\item{par_est}{Estimate of the parameter to be used in deciding on a norm.}

\item{test_stat_func}{A function that will provide the test statistic for the
given fold (using the testing data),
 and uses the best norm (decided on using the training data).}

\item{perf_meas}{The prefered measure to carry out evaluation of how far away the parameter
estimates are away from zero.}

\item{null_quants}{The 95 percent quantiles corresponding to the limiting distribution under the
various norms. The ordering of these quantiles is the same as that of the norm_indx.}

\item{norms_indx}{The index of the norms to be considered.  For example if we use the l_p norm,
norms_indx specifies the different p's to try.}

\item{norm_type}{The type of norm to be used for the test.  Generally the l_p norm}
}
\value{
learned test statistic for a single fold of data
}
\description{
Helper function for cv_test.  This function estimates the cross validated
test statistic for a single fold of data.  Optionally this function can
also ouput the chosen norm index chosen using the training set.
}
