% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noicePaletteGenerator.R
\name{noicePaletteGenerator}
\alias{noicePaletteGenerator}
\title{noicePaletteGenerator}
\usage{
noicePaletteGenerator(ncolor = 1)
}
\arguments{
\item{ncolor}{an integer ranging from 1 to 10.}
}
\value{
a color palette of closest of desired length.
}
\description{
Generate a nice palette (it's more scavenging but well)
}
