% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval_for_mag.R
\name{pval_for_mag}
\alias{pval_for_mag}
\title{Calculate the estimated p-value for a given observed alternative
and a given norm}
\usage{
pval_for_mag(boot_data, dir, lp = 2, nrm_type = "lp")
}
\arguments{
\item{boot_data}{the simulated data draw from
the limiting distribution under the null}

\item{dir}{the observed estimate of the parameter}

\item{lp}{the index for the norm used}

\item{nrm_typ}{the type of norm used}
}
\description{
Calculate the estimated p-value for a given observed alternative
and a given norm
}
