% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Random_Sphere.R
\name{r_unif_sphere}
\alias{r_unif_sphere}
\title{This function is used to generate random observations from the
unit ball (mainly used to generate random directions).}
\usage{
r_unif_sphere(n_covs, n_obs)
}
\description{
This function is used to generate random observations from the
unit ball (mainly used to generate random directions).
}
