% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reviewers_data.R
\docType{data}
\name{reviewers}
\alias{reviewers}
\title{reviewers_data}
\format{
A data frame with X rows and 3 variables:
\describe{
  \item{ Reviewer_1 }{  character }
  \item{ Reviewer_2 }{  character }
  \item{ Reviewer_3 }{  character }
}
}
\source{
Source
}
\usage{
reviewers
}
\description{
Description.
}
\keyword{datasets}
