% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IC-E4.R
\name{rr.msm.ic.4}
\alias{rr.msm.ic.4}
\title{Estimate both the parameter, and the influence
curves used for estimating the projected risk ratio.
These functions will be given an entire dataset, and will
return the estimate for the parameter, and for each
estimated influence curve at each observation. The first column
The first column of obs_data should be the binary outcome of interest.}
\usage{
rr.msm.ic.4(obs_data, what = "both", control = NULL)
}
\arguments{
\item{obs_data}{the observed data.  The first column should be the outcome.}

\item{what}{the desired return value. Should be one of `"ic"`
(infludence curve), `"est"` (estimate), or `"both"`.}

\item{control}{any other control parameters to be passed to the estimator.}
}
\description{
Obtain an estimator of the probability delta = 1 given w
}
