% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IC-DE3.R
\name{rr.msm.ic}
\alias{rr.msm.ic}
\title{Estimate both the parameter, and the influence
curves used for estimating the projected risk ratio.
These functions will be given an entire dataset, and will
return the estimate for the parameter, and for each
estimated influence curve at each observation. The first column
of your data should correspond to the variable of interest
(the variable for which pearson correlation is calculated).}
\usage{
rr.msm.ic(obs_data, what = "both", control = NULL)
}
\arguments{
\item{obs_data}{the observed data.  The first column should be the outcome.}

\item{what}{the desired return value. Should be one of `"ic"`
(infludence curve), `"est"` (estimate), or `"both"`.}

\item{control}{any other control parameters to be passed to the estimator.}
}
\description{
Obtain an estimator of the probability delta = 1 given w
}
