% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IC-DE3jnet.R
\name{rr.msm.jn.ic}
\alias{rr.msm.jn.ic}
\title{Estimate both the parameter, and the influence
curves used for estimating the projected risk ratio from a working
marginal structural mean model.  This function only uses elastic net
for the estimation (rather than also using other learners).}
\usage{
rr.msm.jn.ic(obs_data, what = "both", control = NULL)
}
\arguments{
\item{obs_data}{the observed data.  The first column should be the outcome.}

\item{what}{the desired return value. Should be one of `"ic"`
(infludence curve), `"est"` (estimate), or `"both"`.}

\item{control}{any other control parameters to be passed to the estimator.}
}
\description{
The function takes a dataset, and will returns the estimate for the
parameter, and or the estimated influence curve at each observation.
The first column of obs_data should correspond to the variable of interest
(the variable for which pearson correlation is calculated).
}
\details{
Obtain an estimator of the probability delta = 1 given w
}
