% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_lp.R
\name{select_lp}
\alias{select_lp}
\title{A function used to select which l_p norm to use for
a given test.}
\usage{
select_lp(e_lm_dstr, pos_lp, search_size = 1000, fxd_val = fixed,
  compare_measure = cmp_meas)
}
\arguments{
\item{e_lm_dstr}{is the estimated limiting distribution}

\item{pos_lp}{is a vector containing the possible
values of the lp norm to be considered}

\item{search_size}{how many random draws will be taken to
estimate the average power.}

\item{fxd_val}{which parameter (power or magnitude) you wish
to keep fixed when solvinf for the best lp norm.  If fxd_val
is "power", the criteria used to choose the lp norm will be
based on the magnitude of local alternative required to reach
a specific power.}

\item{compare_measure}{whether you wish to compare the average scinario
verus the worst case scenario.}
}
\description{
A function used to select which l_p norm to use for
a given test.
}
