% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Single_Boostrap.R
\name{single_bootstrap}
\alias{single_bootstrap}
\title{Take a bootstrap draws based on the observed data
and the influence curve.}
\usage{
single_bootstrap(data, ifc, sims, norms = c("max"))
}
\arguments{
\item{data}{The observed data}

\item{ifc}{The influence curve}

\item{sims}{The number of draws desired}

\item{norms}{the norms underwhich the bootstrap draws chould be compressed.}
}
\description{
Take a bootstrap draws based on the observed data
and the influence curve.
}
