% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laber_test.R
\name{ZL}
\alias{ZL}
\title{Carry out a simplified version of the Zhang and Laber test.}
\usage{
ZL(observed_data, ts_sims, ld_sims)
}
\arguments{
\item{observed_data}{The observed data.}

\item{ts_sims}{The number of draws from the test statistic distribution}

\item{ld_sims}{The number of draws from the limiting distribution to estiamte each test statistic.}
}
\description{
Carry out a simplified version of the Zhang and Laber test.
}
