% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laber_test.R
\name{ZL_use_infl}
\alias{ZL_use_infl}
\title{Run a Zhang and Laber test, but use influence functions to obtain estimates of variance
Rather than using the parametric model based variance estimates.}
\usage{
ZL_use_infl(observed_data, ts_sims, ld_sims)
}
\arguments{
\item{observed_data}{The observed data.}

\item{ts_sims}{The number of draws from the test statistic distribution}

\item{ld_sims}{The number of draws from the limiting distribution to estimate each test statistic.}
}
\description{
Run a Zhang and Laber test, but use influence functions to obtain estimates of variance
Rather than using the parametric model based variance estimates.
}
