% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv_pn_test.R
\name{mv_pn_test}
\alias{mv_pn_test}
\title{Runs a multivariate point null test. This function returns
an approximate p-value for the specified test statistic.}
\usage{
mv_pn_test(obs_data, param_est = NULL, control = test.control())
}
\arguments{
\item{obs_data}{The observed data to be used for finding the optimal
norm (training), and finding the test statistic (testing).  Similar to
above, each row is an observation and each column corresponds to either
the outcome (first column) or a covariate.}

\item{param_est}{Function used to estimate the parameter and corresponding
influence curve.}

\item{control}{List used to define controls for test.}
}
\value{
Either a p-value, or a p-value and other information, depending on
the value of control\$more_info.
}
\description{
Runs a multivariate point null test. This function returns
an approximate p-value for the specified test statistic.
}
