% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pow_for_mag.R
\name{pow_for_mag}
\alias{pow_for_mag}
\title{Helper function for est_local_pow.  This function
is used to estimate the power of a
certain direction given the magnitude of that direction.}
\usage{
pow_for_mag(boot_data, dir, lp = 2, mag = 1, nf_quant)
}
\arguments{
\item{boot_data}{bootstrap estimates of the (centered) limiting distribution.
Data where columns correspond to different covariates, and rows are independent observations.}

\item{dir}{a vector in the direction for which we wish to estimate power.}

\item{lp}{the lp norm to be used (ideally it is an integer)}

\item{mag}{length of the local alternative vector.}
}
\value{
Magnitude for the specified lp norm for a given local alternative.
}
\description{
Helper function for est_local_pow.  This function
is used to estimate the power of a
certain direction given the magnitude of that direction.
}
