% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.control.R
\name{test.control}
\alias{test.control}
\title{Control function for the adaptive norm test}
\usage{
test.control(
  n_mc_samples = 300,
  nrm_type = "lp",
  perf_meas = "est_acc",
  pos_lp_norms = c(1, 2, 3, "max"),
  show_hist = FALSE,
  ld_est_meth = "par_boot",
  ts_ld_bs_samp = 250,
  more_info = TRUE,
  ret_cov_mat = FALSE,
  ...
)
}
\arguments{
\item{n_mc_samples}{Number of samples to be used in estimating the limiting
distribution of the test statistic under the null.}

\item{nrm_type}{The type of norm to be used for the test.
Generally the l_p norm}

\item{perf_meas}{the prefered measure used to generate the test statistic.}

\item{pos_lp_norms}{The index of the norms to be considered.  For example if
we use the l_p norm, norms_indx specifies the different p's to try.}

\item{show_hist}{set to True to see a histogram of the test statistic's
limiting distribution MC draws compared to the estimated parameter.}

\item{ld_est_meth}{String indicating method for estimating the limiting
distribution of the test statistic parametric bootstrap or permutation.}

\item{ts_ld_bs_samp}{The number of test statistic limiting distribution
bootstrap samples to be drawn.}

\item{more_info}{Boolean indicating if the function should return more
information that just the p-value.  When true, the chosen norm index,
the bonferroni based p-value, the test statistic,
and the estimated distribution of the test statistic will be returned.}

\item{ret_cov_mat}{Boolean indicating if the function should also return
the estimated covariance matrix.  This can somewhat slow computation and
result in large file sizes at higher dimensions. The argument `more_info`
must be set to `TRUE` for this option to be possible.}

\item{...}{Other arguments needed in other places.}
}
\description{
Control function for the adaptive norm test
}
